requirements.txt
Django==2.2
django-nested-inline==0.4.6
importlib-metadata==4.2.0
mccabe==0.7.0
pycodestyle==2.9.1
pyflakes==2.5.0
python-dotenv==0.21.1
pytz==2022.7.1
sqlparse==0.4.3
typing-extensions==4.5.0
zipp==3.15.0
